Integrate seed dirs; add environment variable
