+++ .dassie/.env
@@ -1,3 +1,6 @@
 MEMCACHED_HOST=memcached
 REDIS_PROVIDER=SIDEKIQ_REDIS_URL
 SIDEKIQ_REDIS_URL=redis://redis:6379/0
+
+# Uncomment next line to enable clearing/seeding the db
+# RAILS_SKIP_SEEDS=false
